@@ -5,10 +5,22 @@ setup(
     version="0.1.0",
     packages=find_packages(where="src"),
     package_dir={"": "src"},
+    python_requires=">=3.12",
     install_requires=[
-        line.strip()
-        for line in open("requirements.txt")
-        if line.strip() and not line.startswith("#")
-    ],
-    python_requires=">=3.9",
+        "fastapi",
+        "pydantic",
+        "pydantic-settings",
+        "pydantic-ai[logfire]",
+        "uvicorn",
+        "python-multipart",
+        "python-dotenv",
+        "logfire",
+        "openai",
+        "httpx",
+        "numpy",
+        "tiktoken",
+        "PyMuPDF",
+        "pdfplumber",
+        "pandas"
+    ]
 )
