@@ -15,7 +15,8 @@ class Settings(BaseSettings):
         UPLOADS_DIR: Directory for uploaded files
         PDF_DIR: Directory for PDF files
         DIAGRAMS_DIR: Directory for diagram files
-        VECTOR_STORE_DIR: Directory for vector store data
+        FRONTEND_PUBLIC_DIR: Directory for frontend public static assets
+        FRONTEND_DIAGRAMS_DIR: Directory for frontend diagram files
     """
     model_config = SettingsConfigDict(
         env_prefix="APP_",
@@ -33,6 +34,11 @@ class Settings(BaseSettings):
     )
 
     @computed_field
+    def PROJECT_ROOT(self) -> Path:
+        """Get the project root directory (one level up from backend)."""
+        return self.BASE_DIR.parent
+
+    @computed_field
     def PDF_DIR(self) -> Path:
         """Get PDF directory and ensure it exists."""
         pdf_dir = self.UPLOADS_DIR / "pdfs"
@@ -42,16 +48,16 @@ class Settings(BaseSettings):
     @computed_field
     def DIAGRAMS_DIR(self) -> Path:
         """Get diagrams directory and ensure it exists."""
-        diagrams_dir = self.UPLOADS_DIR / "diagrams"
+        diagrams_dir = self.FRONTEND_PUBLIC_DIR / "diagrams"
         diagrams_dir.mkdir(parents=True, exist_ok=True)
         return diagrams_dir
 
     @computed_field
-    def VECTOR_STORE_DIR(self) -> Path:
-        """Get vector store directory and ensure it exists."""
-        vector_store_dir = self.UPLOADS_DIR / "vector_store"
-        vector_store_dir.mkdir(parents=True, exist_ok=True)
-        return vector_store_dir
+    def FRONTEND_PUBLIC_DIR(self) -> Path:
+        """Get frontend public directory for static assets."""
+        public_dir = self.PROJECT_ROOT / "frontend" / "public"
+        public_dir.mkdir(parents=True, exist_ok=True)
+        return public_dir
 
     def model_post_init(self, context: Any) -> None:
         """Ensure upload directory exists and is absolute."""
